backend/requirements/base.txt
# Core Django
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# Caching
redis==5.0.1
django-redis==5.4.0

# Celery
celery==5.3.6

# Environment variables
python-dotenv==1.0.1

# API Documentation
drf-spectacular==0.27.1

# Utilities
pytz==2024.1
